requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
rispy
PyMuPDF
openai
